refactor: move workout list and workout form to its own components
